fix long-standing UI bug
This has been annoying me on and off for YEARS!  (Literally.)
In some (quite rare) settings, there was an annoying UI bug
where it would look like some updates were not propagated to
the UI when they had to.

After spending a couple of days investigating this, it turns
out that it is not a bug in the editor, but in xserver (maybe).
Experiments and frustration revealed that the bug seems to be
in some backing store-related logic.  So, here, I remove our
request to get a backing store.

This backing store option was here to help performance, but X
implementations really should be smart enough to use one even
when clients don't ask for it.  Even if X servers are dumb,
removing the option should not cause a noticeable performance
degradation.