implement window resize
Use an offscreen surface for rendering
of the editor's framebuffer, blit it
only on EXPOSE events in the event-
handling thread.

Looks like there are a number of data
races in the SDL2 renderer's code.

1. With accelerated renderer you
   can draw directly only in the
   thread that created the window.

2. If you handle window events in
   a separate thread there are
   races on window resize, only
   coarse-grained locking can help
   in this situation, but use of
   sleeping event wait API totally
   tears down the hopes to resolve
   the problems this way.

Fortunately, rendering to an offscreen
surface saves the situation. The real
blitting is done in the separate thead,
where window's events are handled.